Separated mongoose config into separate file
